+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:arcane/arcane.dart';
-import 'package:example/home.dart';
+import 'package:example/screen/home.dart';
 
 bool v = false;
 String? vv;