+++ example/lib/screen/settings.dart
@@ -1,0 +1,10 @@
+import 'package:arcane/arcane.dart';
+
+class SettingsScreen extends StatelessWidget {
+  const SettingsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) => Screen(
+        title: "Settings",
+      );
+}